Level Generation Integrated
Generating mazes correctly, and drawing them appropriately.
First person camera and movement within the maze are the next phase of
development for the 3d aspects.

+++ Modules/Cheryl/Cheryl/Stat_System.h
@@ -5,12 +5,12 @@
 
 class item;
 
-typedef std::pair<int, int> coords;
+typedef std::pair<short, short> coords;
 
 struct stat_system
 {
 	int strength = 0;				//Damage increases with str
-	int intelligence = 0;			//MDamage increases with intel
+	int intelligence = 0;		//MDamage increases with intel
 	int speed = 0;					//Hit, dodge, etc...
 	int constitution = 0;		//Health, Damage reduction
 	int defense = 0;
@@ -29,14 +29,14 @@ class leveling_system
 class base_game_entity : protected stat_system, protected leveling_system
 {
 private:
-	int movementSpeed = 0;			//movement speed based on spd
-	float dodgeChance = 0;			//dodge chance based on spd
-	int damageReduction = 0;		//Amount of damage you take, reduced by con
+	int movementSpeed = 0;		//movement speed based on spd
+	float dodgeChance = 0;		//dodge chance based on spd
+	int damageReduction = 0;	//Amount of damage you take, reduced by con
 
 	int attackSpeed = 0;				//att speed based on spd and str
 	float hitChance = 0;				//hit chance based on  spd
 	int physSkillDamage = 0;		//Damage done with str based skills
-	int magicSkillDamage = 0;		//Damage done with int based skills
+	int magicSkillDamage = 0;	//Damage done with int based skills
 	int damage = 0;						//Amount of damage you do based on str
 
 	item* weapon = nullptr;
@@ -54,13 +54,13 @@ public:
 	int get_MoveSpeed();					//adjusts movement based on spd
 	float calc_DodgeChance( int Speed_of_Attack );			//adjusts dodge based on spd
 	int get_DamageReduction();		//adjusts dmg taken based on con
-	int get_Health();							//adjusts hp based on con
+	int get_Health();								//adjusts hp based on con
 
 	int get_AttackSpeed();					//adjusts attack speed based on spd and str
 	int get_HitChance();						//adjusts the hit chanc based on spd
 	int get_MagicSkillDamage();		//adjusts magic skill dmg based on intel
 	int get_PhysicalSkillDamage();	//adjusts physical skill dmg based on str
-	int get_Damage();						//adjusts dmg based on str
+	int get_Damage();							//adjusts dmg based on str
 };
 
 #endif STAT_SYSTEM_H
